backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch>=2.0.0
segmentation-models-pytorch==0.3.4
albumentations==1.4.10
opencv-python-headless==4.10.0.84
numpy>=1.24.0
Pillow>=10.0.0
